Restructure and fix re-imports from other components
